Reduce the final number of chunks in the web runtime
This will make sure Webpack doesn't try to generate
new chunk names too aggressively, forcing us to have to
deal with multiple versions for libraries or manually having
to import dynamic chunks.

We can achieve the level of bundle splitting we want by using
named chunk imports (as we do with the runtime).