feat(personas-messaging): add additional context around external ids
As a potential consumer of webhooks for SMS and email it seems necessary that we'd want to know how the phone number or email was chosen to be sent to. Since in the current implementation are grabbing it potentially off the profile we should include the actual external id value since it may not match what Twilio/SendGrid tells us, as well as the profile object key.

GRAMA-283