[STRATCONN-2144] Make error code and status mandatory for Integration Error (#1322)
This PR marks error code and status as mandatory for Integration Error class. Error code is important because they are shown in Event Delivery and Delivery Overview screens and indicates the type of error that has occurred. Status is influential in retries.
Marking them mandatory would enforce builders to spend time providing appropriate error codes and status thereby improving observability.

Errors without error code are tagged as Internal errors and show up in Sentry. This happened for Heap and hence we are taking this effort.

Destinations that weren't providing error codes or status codes have been cleaned up and deployed already as part of different PRs.