Update Attio groupWorkspace to use an ID rather than name
The original plan for the implementation of this object was to ask the
user to specify the name of their workspace (assuming this was unique).
We've since decided based on customer feedback to switch to using a new
ID attribute instead.

This action is not yet deployed anywhere so this change is not breaking
for any customers.