Update Attio action to link User<>Workspace and append values (#1693)
* Update Attio action to possibly capture $.userId in `group` handler

This change allows Attio to link User<>Workspace objects in our system,
finally.

I've implemented it as a non-breaking change because:

  1. We have several existing customers now, include Attio ourselves,
     and I don't want to break the integrations.

  2. The identifyUser / groupWorkspace actions are intended to be usable
     separately, so if you don't want this linking to happen the
     groupWorkspace action should continue to work properly.

* Set ?append_to_existing_values=true when calling Attio Assert API

By default, our Assert Record API will replace any existing multi-select
attributes, so consecutive API calls referring to different Users in the
same Workspace would end up overwriting each other, so a Workspace would
only ever have a single user in it (which write came last).

Since Segment events are supposed to be additive, we've added a special
API flag here to control this behaviour. This change is not breaking for
existing users, and will improve the system.