Make InputFieldJSONSchema.properties work with specialized types
`InputFieldJSONSchema` is a base type that can be specialized. E.g.
`InputField` specializes it.

However, in the base type, `properties` is defined in terms of a
specialized type.

This makes it hard to correctly type `properties`. For example,
specializing `InputField` with, say, `EditableInputField` means that
`properties` would be defined in terms of `InputField`, not
`EditableInputField`.

Instead, we can use TypeScript's [polymorphic this type] to define
`properties`.

[polymorphic this type]: https://stackoverflow.com/a/67307495